+++ CHANGES.rst
@@ -18,6 +18,12 @@ Unreleased (2.6.2) (XXXX-XX-XX)
 
 - Added ng-animate to omnibox.
 
+- Restyle the user menu.
+
+- Restyle the search box.
+
+- Fix bugs in zoom buttons, openLayerGroups and clearing of the search query.
+
 
 Release 2.6.2 (2016-1-18)
 ---------------------